Move utils file to util folder
